requirements.txt
requests
pandas
altair
streamlit
python-dotenv
pytest
pylint